moved eliminateLangFunc to utils
