+++ oarepo_ui/theme/assets/semantic-ui/js/oarepo_ui/forms/components/MultilingualTextInput/MultilingualTextInput.jsx
@@ -8,18 +8,7 @@ import {
   useVocabularyOptions,
 } from "@js/oarepo_ui";
 import { i18next } from "@translations/oarepo_ui/i18next";
-
-const eliminateUsedLanguages = (excludeIndex, languageOptions, fieldArray) => {
-  const currentlySelectedLanguage = fieldArray[excludeIndex].lang;
-  const excludedLanguages = fieldArray.filter(
-    (item) => item.lang !== currentlySelectedLanguage && item.lang
-  );
-  const remainingLanguages = languageOptions.filter(
-    (option) =>
-      !excludedLanguages.map((item) => item.lang).includes(option.value)
-  );
-  return remainingLanguages;
-};
+import { eliminateUsedLanguages } from "@js/oarepo_ui";
 
 export const MultilingualTextInput = ({
   fieldPath,
